requirements.txt
streamlit>=1.42,<2
pypdf>=4.0,<5
openai>=1.40,<2
google-genai>=1.0,<2
requests>=2.31,<3
beautifulsoup4>=4.12,<5
lxml>=5.1,<6
pyyaml>=6.0.1,<7
arxiv>=2.1,<3
kaggle>=1.6.6,<2
python-dotenv>=1.0.1,<2
Pillow>=10.0,<12
plotly>=5.20,<6
pandas>=2.0,<3
